+++ modules/home-manager/services/default.nix
@@ -1,5 +1,6 @@
 {
   imports = [
+    ./gnome-keyring.nix
     ./picom.nix
     ./poweralertd.nix
     ./ssh-agent.nix