+++ modules/nixos/hardware/bluetooth.nix
@@ -12,14 +12,38 @@ in {
 
     experimentalFeatures = mkOption {
       type = types.bool;
-      default = true;
+      default = false;
       description = "Enable experimental Bluetooth features";
     };
 
-    autoConnect = mkOption {
+    powerOnBoot = mkOption {
       type = types.bool;
       default = true;
-      description = "Automatically connect to known devices";
+      description = "Power on Bluetooth adapter on boot";
+    };
+
+    fastConnectable = mkOption {
+      type = types.bool;
+      default = true;
+      description = "Enable FastConnectable for quicker device connections";
+    };
+
+    autoEnable = mkOption {
+      type = types.bool;
+      default = true;
+      description = "Automatically enable Bluetooth adapter";
+    };
+
+    reconnectAttempts = mkOption {
+      type = types.int;
+      default = 7;
+      description = "Number of reconnection attempts";
+    };
+
+    reconnectIntervals = mkOption {
+      type = types.str;
+      default = "1,2,3,4,8";
+      description = "Reconnection interval pattern in seconds";
     };
   };
 
@@ -27,9 +51,15 @@ in {
     hardware.bluetooth = {
       enable = true;
       package = mkIf cfg.experimentalFeatures pkgs.bluez5-experimental;
-      settings = mkIf cfg.autoConnect {
+      powerOnBoot = cfg.powerOnBoot;
+      settings = {
         General = {
-          AutoConnect = true;
+          FastConnectable = cfg.fastConnectable;
+        };
+        Policy = {
+          ReconnectAttempts = cfg.reconnectAttempts;
+          ReconnectIntervals = cfg.reconnectIntervals;
+          AutoEnable = cfg.autoEnable;
         };
       };
     };