+++ modules/nixos/security/default.nix
@@ -1,5 +1,6 @@
 {
   imports = [
     ./doas.nix
+    ./fail2ban.nix
   ];
 }